Improve file names in API
* Name the "main" file in each package after the package name
* Split up types.go into card, board and list files
