+++ api/internal/trello/card.go
@@ -58,40 +58,3 @@ func (jc *jsonCard) Card() Card {
 	card.URL = jc.URL.URL
 	return card
 }
-
-// List represents a Trello list returned via the API
-type List struct {
-	ID   string `json:"id"`
-	Name string `json:"name"`
-}
-
-// Board represents a Trello board returned via the API
-type Board struct {
-	ID          string      `json:"id"`
-	Name        string      `json:"name"`
-	Preferences Preferences `json:"prefs"`
-}
-
-// Preferences represents the preferences for a Trello board returned via the API
-type Preferences struct {
-	BackgroundImages []BackgroundImage `json:"backgroundImageScaled"`
-}
-
-// BackgroundImage represents the background image of a Trello board returned via the API
-type BackgroundImage struct {
-	URL url.URL `json:"-"`
-}
-
-// UnmarshalJSON converts JSON data into a BackgroundImage
-func (b *BackgroundImage) UnmarshalJSON(data []byte) error {
-	var jsonBackgroundImage jsonBackgroundImage
-	if err := json.Unmarshal(data, &jsonBackgroundImage); err != nil {
-		return err
-	}
-	*b = BackgroundImage{jsonBackgroundImage.URL.URL}
-	return nil
-}
-
-type jsonBackgroundImage struct {
-	URL urlWrapper `json:"url"`
-}
